104. Maximum Depth of Binary Tree
