--- src/main/kotlin/maximumDepthOfBinaryTree/MaximumDepthOfBinaryTree.kt
@@ -1,30 +1,0 @@
-package maximumDepthOfBinaryTree
-
-import utils.TreeNode
-import java.util.LinkedList
-import java.util.Queue
-
-class MaximumDepthOfBinaryTree {
-    fun maxDepth(root: TreeNode?): Int {
-        if (root == null) return 0
-        val q: Queue<TreeNode> = LinkedList()
-        q.add(root)
-        var height = 0
-        var cur: TreeNode
-        var size: Int
-        while (q.isNotEmpty()) {
-            height++
-            size = q.size
-            for (i in 0 until size) {
-                cur = q.poll()
-                if (cur.left != null) {
-                    q.add(cur.left)
-                }
-                if (cur.right != null) {
-                    q.add(cur.right)
-                }
-            }
-        }
-        return height
-    }
-}
